requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
absl-py==0.9.0            # via -r requirements.in
altair==4.1.0             # via -r requirements.in
apache-beam==2.22.0       # via -r requirements.in
appnope==0.1.0            # via ipykernel, ipython
attrs==19.3.0             # via jsonschema
avro-python3==1.9.2.1     # via apache-beam
backcall==0.2.0           # via ipython
bleach==3.1.5             # via nbconvert
certifi==2020.4.5.2       # via requests
chardet==3.0.4            # via requests
crcmod==1.7               # via apache-beam
decorator==4.4.2          # via ipython, traitlets
defusedxml==0.6.0         # via nbconvert
dill==0.3.1.1             # via apache-beam
docopt==0.6.2             # via hdfs
entrypoints==0.3          # via altair, nbconvert
fastavro==0.23.4          # via apache-beam
future==0.18.2            # via apache-beam
grpcio==1.29.0            # via apache-beam
hdfs==2.5.8               # via apache-beam
httplib2==0.17.4          # via apache-beam, oauth2client
idna==2.9                 # via requests
ipykernel==5.3.0          # via ipywidgets, notebook
ipython-genutils==0.2.0   # via nbformat, notebook, traitlets
ipython==7.15.0           # via ipykernel, ipywidgets
ipywidgets==7.5.1         # via -r requirements.in
jedi==0.17.0              # via ipython
jinja2==2.11.2            # via altair, nbconvert, notebook
jsonschema==3.2.0         # via altair, nbformat
jupyter-client==6.1.3     # via ipykernel, notebook
jupyter-core==4.6.3       # via jupyter-client, nbconvert, nbformat, notebook
jupyter-http-over-ws==0.0.8  # via -r requirements.in
markupsafe==1.1.1         # via jinja2
mistune==0.8.4            # via nbconvert
mock==2.0.0               # via apache-beam
mongomock==3.19.0         # via -r requirements.in
nbconvert==5.6.1          # via notebook
nbformat==5.0.7           # via ipywidgets, nbconvert, notebook
notebook==6.0.3           # via -r requirements.in, jupyter-http-over-ws, widgetsnbextension
numpy==1.18.5             # via altair, apache-beam, pandas, pyarrow
oauth2client==3.0.0       # via apache-beam
packaging==20.4           # via bleach
pandas==1.0.4             # via -r requirements.in, altair
pandocfilters==1.4.2      # via nbconvert
parso==0.7.0              # via jedi
pbr==5.4.5                # via mock
pendulum==2.1.0           # via -r requirements.in
pexpect==4.8.0            # via ipython
pickleshare==0.7.5        # via ipython
prometheus-client==0.8.0  # via notebook
prompt-toolkit==3.0.5     # via ipython
proto-matcher==0.0.2      # via -r requirements.in
protobuf==3.12.2          # via -r requirements.in, apache-beam
ptyprocess==0.6.0         # via pexpect, terminado
pyarrow==0.17.1           # via apache-beam
pyasn1-modules==0.2.8     # via oauth2client
pyasn1==0.4.8             # via oauth2client, pyasn1-modules, rsa
pydot==1.4.1              # via apache-beam
pygments==2.6.1           # via ipython, nbconvert
pyhamcrest==2.0.2         # via -r requirements.in
pymongo==3.10.1           # via -r requirements.in, apache-beam
pyparsing==2.4.7          # via packaging, pydot
pyrsistent==0.16.0        # via jsonschema
python-dateutil==2.8.1    # via apache-beam, jupyter-client, pandas, pendulum
pytz==2020.1              # via apache-beam, fastavro, pandas
pytzdata==2019.3          # via pendulum
pyzmq==19.0.1             # via jupyter-client, notebook
requests==2.23.0          # via -r requirements.in, hdfs
rsa==4.6                  # via oauth2client
send2trash==1.5.0         # via notebook
sentinels==1.0.0          # via mongomock
six==1.15.0               # via absl-py, bleach, grpcio, hdfs, jsonschema, jupyter-http-over-ws, mock, mongomock, oauth2client, packaging, protobuf, pyrsistent, python-dateutil, traitlets
terminado==0.8.3          # via notebook
testpath==0.4.4           # via nbconvert
toolz==0.10.0             # via altair
tornado==6.0.4            # via ipykernel, jupyter-client, jupyter-http-over-ws, notebook, terminado
traitlets==4.3.3          # via ipykernel, ipython, ipywidgets, jupyter-client, jupyter-core, nbconvert, nbformat, notebook
typing-extensions==3.7.4.2  # via apache-beam
urllib3==1.25.9           # via requests
wcwidth==0.2.4            # via prompt-toolkit
webencodings==0.5.1       # via bleach
widgetsnbextension==3.5.1  # via ipywidgets

# The following packages are considered to be unsafe in a requirements file:
# setuptools
